Refactor energy estimation into extensible module
Replace monolithic EnergyConfig/EnergyModel with pluggable architecture:
- Add EnergyEstimator abstract interface for energy estimation
- Add IRBasedEstimator as concrete implementation
- Add EnergyEstimatorFactory with type registration for extensibility
- Require explicit estimator_type in config to prevent errors
- Update CFGAnalysis to use estimator interface (energyCost now double)
- Delete legacy EnergyConfig.h/cpp and EnergyModel.h/cpp
- Document extensibility pattern in README

+++ passes/src/CheckpointOptimizer.cpp
@@ -106,7 +106,7 @@ void CheckpointOptimizer::addObjective() {
 
 void CheckpointOptimizer::addConstraints() {
     // Compute Big-M constant
-    int maxEnergy = 0;
+    double maxEnergy = 0.0;
     for (const auto &blockName : cfg_.getBlocks()) {
         const BlockInfo &info = cfg_.getBlockInfo(blockName);
         maxEnergy = std::max(maxEnergy, info.energyCost);
